+++ lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_screenshot_helper.rb
@@ -4,6 +4,9 @@ require 'onlyoffice_s3_wrapper'
 module OnlyofficeWebdriverWrapper
   # Working with screenshots
   module WebdriverScreenshotHelper
+    # @return [String] content type of png image
+    SCREENSHOT_CONTENT_TYPE = 'image/png'
+
     # @return [OnlyofficeS3Wrapper::AmazonS3Wrapper] s3 wrapper
     def amazon_s3_wrapper
       @amazon_s3_wrapper ||= OnlyofficeS3Wrapper::AmazonS3Wrapper.new(bucket_name: 'webdriver-wrapper-screenshots',
@@ -18,7 +21,7 @@ module OnlyofficeWebdriverWrapper
     def get_screenshot_and_upload(path_to_screenshot = "#{screenshot_folder}/#{StringHelper.generate_random_string}.png")
       begin
         get_screenshot(path_to_screenshot)
-        cloud_screenshot = amazon_s3_wrapper.upload_file_and_make_public(path_to_screenshot)
+        cloud_screenshot = publish_screenshot(path_to_screenshot)
         File.delete(path_to_screenshot) if File.exist?(path_to_screenshot)
         OnlyofficeLoggerHelper.log("upload screenshot: #{cloud_screenshot}")
         return cloud_screenshot
@@ -49,20 +52,31 @@ module OnlyofficeWebdriverWrapper
         if @headless.headless_instance.nil?
           system_screenshot("/tmp/#{screenshot_name}.png")
           begin
-            link = amazon_s3_wrapper.upload_file_and_make_public("/tmp/#{screenshot_name}.png")
+            link = publish_screenshot("/tmp/#{screenshot_name}.png")
           rescue Exception => e
             OnlyofficeLoggerHelper.log("Error in get screenshot: #{e}. System screenshot #{link}")
           end
         else
           @headless.take_screenshot("/tmp/#{screenshot_name}.png")
           begin
-            link = amazon_s3_wrapper.upload_file_and_make_public("/tmp/#{screenshot_name}.png")
+            link = publish_screenshot("/tmp/#{screenshot_name}.png")
           rescue Exception => e
             OnlyofficeLoggerHelper.log("Error in get screenshot: #{e}. Headless screenshot #{link}")
           end
         end
       end
       "screenshot: #{link}"
+    end
+
+    private
+
+    # Publish screenshot
+    # @param [String] path to file
+    # @return [String] publis internet link to file
+    def publish_screenshot(path)
+      amazon_s3_wrapper.upload_file_and_make_public(path,
+                                                    nil,
+                                                    SCREENSHOT_CONTENT_TYPE)
     end
   end
 end